Change all iconUri types to CharacterIconSet
Changed all iconUris types to CharacterIconSet which includes all the
possible types of icons for a character. This allows developers to
target individual icon uri's in code.

TIL: Model binding in observables remove any function declarations in
the resulting model object. Trying to create an asArray function in the
model class would result in a null variable declaration. Changing it to
a static class method removed the ability for the method reference to
be nulled.