Implement barebones formation feature
Implemented a barebones version of the formation creation component.
A DisplayEntityComponent was created to be a low level component that
displays entities that implement the DisplayableEntity interface.

A TwoDArray class was made to handle two dimensional array
manipulations.

Entity Framework Core will assign ids for objects so factory id
assignments are wrong.