Implement CombatPanelComponent in the front end
The IconDisplayComponent is now the lowest level display component for
icons to allow displaying content other than CharacterIconSets.

Implemented the CombatPanelComponent in the front end, although it still
needs a lot of tweaks before it is in satisfactory condition.

Buttons should be laid out vertically and needs some css to style.
The targetsGrid should be rotated and both grids should be displayed
side to side.
Clicking a command button should clear the activeAbility and
activeCategory but there is a bug that will prevent switching to the
category panel.
Font colors in the panel should be changed to a clearer more bold color.