Fix issues with formation and combatPanel
The FormationComponent now correctly functions, allowing the user to
create and update new Formations. There are now functionality to set the
name of the Formation and the leaders of each Formation.

The formation grids in the changed components are now fixed to correctly
display a formation. In many cases, whereas the correct implementation
of a 2d array of a grid would have rows in the first array set and
columns in the second array set, their implementations would be
backwards, which caused many errors in display.

Null checking in javascript is slowly starting to become a problem,
causing significant bloat in code.