Implement combat on the game side of the app
Created many new entity classes to hold combat data.

Many abstractions in the form of Factories and interfaces were
implemented for many of the manager classes.

The MapManager was changed to allow multiple WorldEntities to stack on
top of one another.

The current implementation of combat makes a PlayerEntityManager store
the BattleManager reference.