Attempt to fix game bugs
Attempted to fix the bugs that plagued the GameComponent. There is
some sort of connectivity issue that prevents the WorldEntityService
from successfully connecting to and receiving data from the
WorldEntityHub from the server. Ultimately decided that the issue is too
deeply rooted in the code and so a rewrite of the affected classes will
be in order.