Fix game bug that prevents entities from loading
Fixed the bug that prevented entities from loading into the world map.
Very many bugs in the GameComponent were caused by an infinite loop in
the animateEntities function. There were also many edge cases that
prevented the component from correctly setting the isAnimating variable
to false which contributed to the infinite loop.

Changed the recursive loop of animateEntities to use a setTimeout of
0ms. This will free the thread to perform work on other tasks before
returning to work on animating more entities. This will stop the Ui from
freezing while the thread is busy animating.