Fix game world reconnect bug
Fixed a bug where switching from the GameComponent to another component
and back again would cause the client to stop receiving SignalR
messages.

The bug was caused from an exception that was being thrown in the
StartConnection function in the WorldEntityHub due to the
PlayerEntityManager's Entity property being null. Whenever the client
disconnects from the WorldEntityHub, the created PlayerEntityManager is
destroyed and when reconnecting, is recreated. The problem is that the
recreated PlayerEntityManager never gets its Entity property set.