Add WorldEntity name colors and fix map bug
Added the ability for the client to differentiate between player
WorldEntities, enemy WorldEntities, and other players' WorldEntities.
This allows the client to color code the names of the entities to more
effectively visually represent them.

Fixed a bug where reloading the page on the GameComponent while on a
different map than the starting map will load the starting map.