Increase performance of animateEntities loop
Increased the performance of the animateEntities loop in the
GameComponent by changing the forEach loop origin from entityLocations
to tileNodeComponents. This way, there won't be multiple searches for
the oldLocation and the tileNodeComponent that holds the target entity.

The speed of the loop could have been increased further if the
QueryList object that holds the tileNodeComponents allowed index lookups
so that we can skip more searches through arrays. Using the toArray
function from the QueryList object only partially solves this problem as
the array returned does not track changes to its elements. So those
changes would not translate to the ui.