Decouple formation child components from parent
Decoupled the FormationComponent's child components from knowing the
implementation of the parent. Changed the child components from storing
the callback function provided by the parent component to simply
emitting values to the parent instead. This drastically increases code
reusablility and allows the usage of the formationGrid component in the
battle component with more ease.

Most uses of the DisplayableEntity interface particularly in the
FormationComponent and its child components have been removed. The
initial purpose of the interface was to allow other components that
aren't concerned with whether the input was from WorldEntities or
CombatEntities use them. But the use cases of those forementioned
components are now specifically for battle-related purposes only.