Create BattleService and hook up backend to front
Created the BattleService and hooked it up to the backend to receive
battle events. The BattleService is different from other services
created in the past as it uses ReplaySubjects to emit values to any
listening components instead of storing callbacks to invoke as
messages are received from the server.

Added death detection to the BattleManager to implement the EndOfBattle
event. If resurrection is ever added to the game, must manually
increment the number of living CombatEntities in battle.