Change BattleManager to invoke SuccessfulAction
Changed the BattleManager to invoke the SuccessfulAction event instead
of returning all of the CombatEntities affected by a successful action.
This was due to realizing that clients other than the one performing
the action would not receive the results of that action because the
BattleHub is the only place sending the results. Instead, the
BattleHubContext will send the data received by the invoked
SuccessfulAction event and send it to all concerned clients.