Implement majority of battle feature
Most of the battle feature has been implemented both on the front end
and backend.

However, the current method of getting into combat is very flawed. The
flaws are strongly tied to the way WorldEntities move across the world
map, namely "bumping" into an enemy entity allows a player to chooose to
start combat. There are many problems with this method, but fixing it
means changing the methods of moving entities for players.