Add context menu and fix clientside bugs
Added a ContextMenuComponent that barely works and will be used to
perform actions on a target location or WorldEntities.

Added length properties to the Dictionary and PriorityQueue classes.

Fixed a bug with isValidLocation in the Pathfinder class where the
function incorrectly checks the bounds of the map.

Fixed a bug where the Pathfinder would store nodes with higher gScores
as a higher priority node. This is incorrect, as the resulting path
would be the path with the highest gScore, ie the longest path. The fix
was to switch the priority function to make lower gScores the higher
priority nodes.