Change create and join battle methods
Changes the ways of creating and joining battles. Now with pathfinding
added, the PlayerEntityManager decides when it should initiate or join
battle, which is at the end of a move queue order. BattleManagers are
now moved from being managed by PlayerEntityManager instances to being
managed by instances of MapBattleManagers which are created for every
map. This allows the creation of player parties which can all join in
battles.

The AiEntityManager will now attempt to create battles if it walks into
a position with a player WorldEntity residing. Will add joining fellow
ai WorldEntities in a later update.

Created a MovementManager service for use in the GameComponent that is
used to move the player's WorldEntity and keep track of and
automatically move to moving WorldEntities in cases where the player
wants to attack or join an entity.