Add null checking for MapManager functions
Added null checking for MapManager functions that require a WorldEntity
to be added to the map before they should logically work, such as
moving or removing a WorldEntity from the map.

Changed the IMapManager interface to only require one set of
functions instead of making a different function for ai usage and
another for player usage.