Implement initiate and join battle for ai
Implemented the ability for AiEntityManagers to initiate and join
battles. The AiEntityManagers will prioritize joining the battles of its
allies before initiating a new battle with one of its enemies.

Who the AiEntityManager considers an ally or enemy is determined and can
be customized in the TryBattle method, allowing the potential for
AiEntityManagers who are allies with human players and enemies with
other AiEntityManagers.