Implement basic combat ai
Implemented a basic combat ai that uses a threat system to choose which
targets to act on. There's a large amount of balancing that must be done
to ensure the ai makes decisions that are enjoyable to play against and
with.

The current implementation uses an AiRandomnessFactor that makes an ai
behave a little more irrationally. This can be set per enemy type,
allowing some enemies to be more intelligent and others to be less
predictable.

The ai has not yet been added to the BattleManager.