Fix stackoverflow issue
Previously, == operator overload in the Coordinate class did not check
for null equality. The fix for it checked for null equality using the ==
operator which ended up causing an infinite loop from a Coordinate
object recursively calling the == operator comparison.

The fix was to null check without using said == operator and instead
use a simple 'is null' check.