Fix player movement on world map
Fixed the player movement on the world map, allowing the player to right
click an empty tile node to move the player's character there. Occupied
tiles will create a context menu which allows the user to select from a
list of actions to perform on that tile or the characters on that tile.

The context menu is a bit buggy at the moment, showing in the wrong spot
and only if the browser is not fullscreen.