Fix battle initiation
Fixed the battle initiation from the WorldMap for both the ai and the
players. Now the client flows from the GameComponent to the
BattleComponent successfully, although more work is needed to complete
the BattleComponent.

Attempting to start a new battle with a target ai WorldEntity will
include all ai WorldEntities occupying the same coordinate as the target
in the new battle. Trying to start a battle with a player WorldEntity
will bring all of the target player's allies into battle.