Fix bugs with battle start
Added IPlayerEntityManagerStore to the StartupExtensions dependency
injection list to fix the instantiation of BattleHub objects.

Removed IReadOnlyDictionary from an IReadOnlyBattle interface because
the compiler could not cast a dictionary of mutable objects to a
read-only dictionary of immutable objects.

Removed the ability for the client to target the player's own
WorldEntity as the target of an action.

Prevents the client from calling EndConnection on the WorldEntityHub if
the player is in a battle. What this does is it prevents the GC from
cleaning up the old PlayerEntityManager which will allow the player's
WorldEntity to remain in the world map while in battle. This also fixes
an issue where a PlayerEntityManager gets garbage collected before
getting used again in the BattleComponent, losing data crucial for the
BattleComponent.