Add game tick functionality to BattleManagers
Added an OnSecondElapsed function to BattleManagers and an OnGameTick
function to MapBattleManagers. This allows BattleManagers to call its
own functions based on time, which can later be used to allow the Ai
to take its turns in battle in an orderly fashion instead of all at
once.

There is a mismatch between the backend turn time elapsed and the front
end's turn time elapsed. The front end is about 5 seconds faster.