Fix world map movement bugs
Fixed a bug that prevents the player's character from moving after
refreshing a page if no other character moves. This was caused by
ineffectively retrieving the player's WorldEntity id.

Fixed a bug that prevented a new Map from loading when changing to
another map. This was caused from subscriptions to the onChangeMaps
subject not calling a binded version of the changeMapsHandler.

Fixed a bug where every map switch progressively increased the movement
speed of the player's character. This was caused because every change
maps request would subscribe to the OnGameTick event twice in the
PlayerEntityManager and only unsubscribe once.