Fix bugs with battle component
Fixed a bug with the timer component where after counting down to 0, the
timer will stay at 0 even in new turns.

Fixed the PlayerManagerListener and BattleListener to correctly listen
to all related events, which fixed the problem of updates from the
server failing to reach the client.

Added a GameStateAttribute to the server to check for valid states
before allowing a request to reach an action on a controller.