Fix Create and Formation components
Fixed a bug that prevented the user from editing a previously existing
CombatEntity.

Changed the UpdateAsync method in the CombatEntityFactory to return the
same CombatEntity instance that is being updated. This allows the
changes to that CombatEntity to propagate to all Formations that
reference that entity.

Fixed Formation editing in the FormationComponent.