Fix CombatEntity editing in CreateComponent
There was a bug where editing a CombatEntity in the CreateComponent
would not translate those changes to the Formation that CombatEntity
exists in. To fix it, the CombatEntityFactory no longer creates and
replaces CombatEntities when updating them, instead it simply updates
the values of the CombatEntity being updated. This allows the reference
that a Formation object holds of the CombatEntity to remain the same
even in the event that that CombatEntity has its values altered.

One problem that has become apparent in that there has to be a
coordination between the changes to a CombatEntity's iconUris and the
WorldEntity that represents that CombatEntity on the world map. Simply
holding a reference to that iconUris object is not enough to allow any
changes to propagate over to the WorldEntity. This is important when
dealing with changing equipment later on.