Change styling for GameComponent and children
Changed the css styling for the GameComponent, ContextMenuComponent, and
the ContextMenuListComponent to more closely resemble the finished
design.

Prevented event bubbling from opening a context menu on the name part of
a WorldEntityComponent.

Fixed the autoscrolling mechanism in the ContextMenuComponent. The
current implementation moves the scrollbar erratically because of the
change detection inherent in Angular. Moving the scrollbar in the same
function as adding a new MessageComponent will move it before the DOM
renders the new component, retrieving the wrong position data.