Add locks to BattleManager
Added locks to the BattleManager which needs thread synchronization very
much. This will allow all commands flowing from the clientside to
funnel in one at a time, preventing concurrency issues.

Added in a mechanism for telling if a battle is still active, allowing
the BattleManager to discard any commands after the battle has ended and
preventing it from further altering any data afterwards.