Add styling for dead CombatEntities
Added an opacity style for CombatEntities that are dead so players can
more easily differentiate between dead and living CombatEntities.

Adding that meant adding a resources property to the DisplayableEntity
interface which will allow us to display other resources later on. It
also means removing the implementation of DisplayableEntity from the
WorldEntity model object.