Fix successful action display bug
Fixed a bug where the BattleComponent failed to correctly display the
changes to CombatEntities from a successful action.

Combat now flows from start to end of battle correctly.

Sending a request for a BattleAction in the BattleHub will set the
LastAccessed property in the PlayerEntityManager, preventing a player's
manager from being garbage collected for inactive activity while in
battle.