Fix minor bugs with battle
Fixed a bug where a player who joined a battle will be stuck in a battle
indefinitely even when the battle has been concluded.

Fixed a bug where an ai's WorldEntity is not removed from the game map
when all of its CombatEntities are killed in battle.

Fixed a bug where it was impossible to hit targets hiding behind dead
ally CombatEntities.

Changed the StateHandlerService to handle the 'free' state by routing
the client to the GameComponent.