Change StartTurn timer to System.Timer
Changed the timer that calls StartTurn from a System.Threading.Timer to
a System.Timer.Timer. This is because System.Threading.Timer, despite
its name, is not suitable for uses in multithreaded environments such as
the BattleManager in which it is used. System.Timer.Timer is built for
multithreading.

This change removed the deadlocks that occurred when trying to call
StartTurn from the timer.