Fix negative damage bug
Fixed a bug where deducting armor values that are higher than a given
damage value would result in a negative damage value that would heal
the CombatEntity instead of dealing no damage. The proper implementation
is to deal no damage of that particular type if that armor value is
higher. Only in the case of damage resistances can damage be negative
and heal a CombatEntity.