Implement apply StatusEffects at start of turn
Implemented applying damage and healing from StatusEffects at the start
of a turn in battle. Also implemented the removal of those StatusEffects
when the duration counter hits 0.

Fixed a bug that allowed a CombatEntity to be healed past 100% health.

Changed the List of Tasks in the WorldState game loop to a
ConcurrentBag. Because new Tasks were being added in a parallel manner,
a List is dangerous to add to in a multi-threaded manner.