Fix battle-related bugs
Fixed a bug where an AiEntityManager will continue to move and join and
create battles after joining a battle.

Fixed a bug where a death in battle will cause the BattleManager to
decrement living CombatEntities for the wrong team.

Fixed a null-reference exception that occurs at the start of battle
because no ActionPointData is sent with the event.