Adjust BattleComponent user interface
Wrapped all BattleComponent hud elements in a sticky 'navbar' to allow
players to see important data even when they are scrolled down.

Allowed players to have active entities even when it is not their turn.
This fixes a minor bug where a player who reloaded or was joining a
battle when it is not their turn would have no active entity.

The AbilityPanel now makes an active ability have a green background
allowing the player to see which ability is selected. Fixed the issue
with the UI where the CombatPanel would repeatedly expand and shrink
because of the mouseenter and mouseleave events.

At the end of a turn and at the start of a player's turn, the
CombatPanel now correctly resets the currently selected ability for the
player.