Fix crash when attacking a joiner
Fixed a bug which crashed the server when a player attempted to attack
an AI WorldEntity that joined the battle of another AI WorldEntity. This
was because any WorldEntity joining a battle would not get it's id
registered with the MapBattleManager and so the MapBattleManager
wouldn't see that entity as in battle.