Activate combat ai
Activated the combat ai which now functionally attacks and defends in
battle. The threat system has not been implemented yet so the ai simply
chooses the targets the it will deal the most damage to.

The Defend command now works as intended for both the ai and player
usage.

Fixed a bug where an Ai WorldEntity would fail to be removed from combat
if victorious against a player.

Fixed several minor Ui related bugs in the BattleComponent.